Editing the Landing page.
(HTML)
Adding a dummy lorem ipsum text to check functionality of the nav-bar.
Creating a container.
(CSS)
Setting up z-index on the nav-bar to be greater than the values for other sections.
Center for the input section and button.
Adding a display flex: direction to column.
Padding for the container.
(Assets)
Creating an image folder with images for the carousel.